pyproject.toml
[project]
name = "mentorflow"
version = "0.1.0"
description = ""
authors = [
    {name = "KXNY4",email = "shakhdijarergali@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <4.0"
dependencies = [
    "django-redis (>=6.0.0,<7.0.0)",
    "pytest (>=8.4.1,<9.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "autoflake (>=2.3.1,<3.0.0)",
    "django-autoslug (>=1.9.9,<2.0.0)",
    "django-celery-results (>=2.6.0,<3.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "ipython (>=9.4.0,<10.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "django-debug-toolbar (>=6.0.0,<7.0.0)",
    "djoser (>=2.3.3,<3.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-celery-beat (>=2.8.1,<3.0.0)",
    "flower (>=2.0.1,<3.0.0)",
    "django-extensions (>=4.1,<5.0)",
    "django-celery-email (>=3.0.0,<4.0.0)",
    "django-model-utils (>=5.0.0,<6.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.3.0"
black = "^25.1.0"

